Safari View implemented. Fixed a bug on iPad that would cause the title of the VC to not update when moving to a different subreddit.
