+++ RedditRealTime/RedditRealTime/Controller Layer/Table View Cells/SubredditTableViewCell.swift
